+++ src/test/java/com/github/klousiaj/junit/NginxIntegrationTest.java
@@ -1,22 +1,32 @@
 package com.github.klousiaj.junit;
 
+import com.spotify.docker.client.messages.Container;
 import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Test;
 
 import java.io.BufferedInputStream;
+import java.util.List;
 
 /**
  * Created by klousiaj on 8/11/16.
  */
 public class NginxIntegrationTest {
 
+  static String containerName = "leaping-allison";
+  static String[] labels = {"com.github.klousiaj.creator:not-docker-junit-rule",
+    "com.github.klousiaj.test-class:NginxIntegrationTest",
+    "com.github.klousiaj.version:latest"};
+
   @ClassRule
   public static DockerRule nginxRule =
     DockerRule.builder()
       .image("kitematic/hello-world-nginx:latest")
+      .containerName(containerName)
       .ports("80")
       .waitForPort("80/tcp", 8000)
+      .labels(labels)
+      .cleanVolumes(true)
       .build();
 
   @Test
@@ -37,5 +47,25 @@ public class NginxIntegrationTest {
       "    <p>To edit files, double click the <strong>website_files</strong> folder in Kitematic and edit the <strong>index.html</strong> file.</p>\n" +
       "  </div>\n" +
       "</div>\n", output);
+  }
+
+  @Test
+  public void validateContainerName() throws Exception {
+    Assert.assertEquals(containerName, nginxRule.getContainerName());
+  }
+
+  @Test
+  public void validateContainerLabel() throws Exception {
+    List<Container> containers = nginxRule.getContainerByLabel("com.github.klousiaj.creator", "docker-junit-rule");
+    Assert.assertNotNull(containers);
+    Assert.assertEquals(1, containers.size());
+
+    containers = nginxRule.getContainerByLabel("com.github.klousiaj.creator", "not-docker-junit-rule");
+    Assert.assertNotNull(containers);
+    Assert.assertEquals(0, containers.size());
+
+    containers = nginxRule.getContainerByLabel("com.github.klousiaj.test-class");
+    Assert.assertNotNull(containers);
+    Assert.assertEquals(1, containers.size());
   }
 }