+++ src/test/java/com/github/klousiaj/junit/RabbitIntegrationTest.java
@@ -10,6 +10,7 @@ import org.junit.Test;
 
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Pattern;
 
 public class RabbitIntegrationTest {
 
@@ -23,6 +24,7 @@ public class RabbitIntegrationTest {
       .waitForLog("Server startup complete")
       .useRunning(true)
       .leaveRunning(true)
+      .cleanVolumes(true)
       .build();
 
   @Test
@@ -68,11 +70,11 @@ public class RabbitIntegrationTest {
     int preCount = preStartContainers.size();
     int preRunningCount = 0;
     int preRabbitCount = 0;
-    for(Container container : preStartContainers){
-      if("running".equals(container.state())){
+    for (Container container : preStartContainers) {
+      if ("running".equals(container.state())) {
         preRunningCount++;
       }
-      if("rabbitmq:management".equals(container.image())){
+      if ("rabbitmq:management".equals(container.image())) {
         preRabbitCount++;
       }
     }
@@ -105,11 +107,11 @@ public class RabbitIntegrationTest {
     int postCount = postStartContainers.size();
     int postRunningCount = 0;
     int postRabbitCount = 0;
-    for(Container container : postStartContainers){
-      if("running".equals(container.state())){
+    for (Container container : postStartContainers) {
+      if ("running".equals(container.state())) {
         postRunningCount++;
       }
-      if("rabbitmq:management".equals(container.image())){
+      if ("rabbitmq:management".equals(container.image())) {
         postRabbitCount++;
       }
     }
@@ -117,5 +119,11 @@ public class RabbitIntegrationTest {
     Assert.assertEquals(preCount, postCount);
     Assert.assertEquals(preRunningCount, postRunningCount);
     Assert.assertEquals(preRabbitCount, postRabbitCount);
+  }
+
+  @Test
+  public void validateContainerName() throws Exception {
+    Pattern validForDocker = Pattern.compile(DockerRule.CONTAINER_NAME_REGEX);
+    Assert.assertTrue(validForDocker.matcher(rabbitRule.getContainerName()).matches());
   }
 }